net/octeontx_ep: add parameter to select SDP packet mode
Add the 'sdp_packet_mode' parameter to allow the EP driver
to be used in both nic mode, with 24/8 byte headers to/from OcteonTX,
and in loop mode with no headers.  When used in loop mode unmodified
DPDK applications such as testpmd can be used on both the host
and OcteonTX.  If the parameter is not provided, the driver defaults
to nic mode to match the prior behavior.
Usage: sdp_packet_mode=<nic|loop>

Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: I29da2d806e2d24e4f3a46c5103372a8ca1fdcc0b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/58739
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/59046
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/67648