drivers: add support for switch header type pre L2
Adding changes to configure switch header type pre L2 for cnxk.
Along with switch header type user needs to provide the
offset with in the custom header that holds the size of the
custom header and mask for the size with in the size offset.

ci: skip_roc_check skip_checkformat

Signed-off-by: Kiran Kumar K <kirankumark@marvell.com>
Change-Id: I1f114570c6ba451a8d57c27715cb4c5b65d5b2cb
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/64820
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Satheesh Paul <psatheesh@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/67029