net/cnxk: handle external buffers
Handle external buffers in TX path. This implementation is deviating
from the spec in the following way.

The implementation calls the free callback immediately during the
transmit routine as part of the freeing of the mbuf to which the
external buffer is attached to. Hence the free callback that is
passed to the external buffer must always be a no op to ensure that
the external buffer is not freed before transmit.

The actual freeing of the external buffers must be handled by the
user application.

Signed-off-by: Ashwin Sekhar T K <asekhar@marvell.com>
Change-Id: Ib60431de2cc8aabdb80a1e7abd13e041fff77204
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/69142
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>