common/cnxk: validate length argument
The x->len is passed as argument to npc_prep_mcam_ldata(). In the
function the len is used to reference elements of int_info and
int_info_mask array. The arrays are of length NPC_MAX_EXTRACT_DATA_LEN.

Validating the x->len value so that it is not greater than
NPC_MAX_EXTRACT_DATA_LEN.

This patch also resolves warning observed with gcc 12 compiler.

log:
   ../drivers/common/cnxk/roc_npc_utils.c:13:26: warning: writing 16 bytes
   into a region of size 0 [-Wstringop-overflow=]
   ptr[idx] = data[len - 1 - idx];

   ../drivers/common/cnxk/roc_npc_utils.c:163:17: note: at offset 64 into
   destination object ‘int_info’ of size 64
   uint8_t int_info[NPC_MAX_EXTRACT_DATA_LEN];

Bugzilla ID: 854
Fixes: 665b6a7400bf ("common/cnxk: add NPC helper API")

Signed-off-by: Ankur Dwivedi <adwivedi@marvell.com>
Change-Id: I13fe0aaadf2249fd8365e90d2500d461ea651d08
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/70368
Reviewed-by: Kiran Kumar Kokkilagadda <kirankumark@marvell.com>
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>