common/cnxk: support for CPT second pass flow rules
Added support to create flow rules to match packets
from CPT(second pass packets). With this change, ingress
rules will be created with bits 10 and 11 of channel field
in the MCAM ignored by default. For rules specific to
second pass packets, the CPT channel bits will be set
in the MCAM.

Signed-off-by: Satheesh Paul <psatheesh@marvell.com>
Change-Id: I7ca8c5ecd67a8a1ac1e30d75e1168d94ea0a7631
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/71477
Reviewed-by: Kiran Kumar Kokkilagadda <kirankumark@marvell.com>
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>