ci: rerun failed asim tests always and remove known issue checks
Currently when an ASIM test fails, the known issue status is checked and
then only it is decided whether to re-run the test or not.

Since ASIM testing has been made as nightly runs, the time saved from this
micro optimization is not very meaningful now. So simplify this logic to
re-run failed test always until ASIM_MAX_REBOOTS count is hit.

With this change the known issue checks can be completely removed.

ci: nightly_test-asim-cn10ka

Signed-off-by: Ashwin Sekhar T K <asekhar@marvell.com>
Change-Id: Ibbc5692c0463433915e79b8aa689285ad4154abc
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/71556