common/cnxk: add multi channel support for SDP send queues
Currently only base channel number is configured as default
channel for all the SDP send queues. Due to this, packets
sent on different SQ's are landing on the same output queue
on the host. Channel number in the send queue should be
configured according to the number of queues assigned to the
SDP PF or VF device.

Signed-off-by: Subrahmanyam Nilla <snilla@marvell.com>
Change-Id: I63236d9e5b26fb8060ff5cc7a9aeeab33975c549
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/72885
Reviewed-by: Radha Chintakuntla <radhac@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>
Tested-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>