event/cnxk: fix overwriting LMT buffer
When transmitting an event vector if group of packets are
not destined to the same queue then they are transmitted
one by one using the same LMT line to a target SQ.
The LMT is pushed to the HW using STOER instruction and
we have to make sure that the STOER has completed before
we re-use the LMT line.
Add a write barrier to make sure that the STOER completed
before reusing the LMT line.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Change-Id: I973c71cf520d0cc3e640ec20c51213e5aeba7a37
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/84029
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>