app/crypto-perf: fix IPsec direction
The default value of options->auth_op & options->cipher_op are such that
an unconditional check for the same would always return true. Hence, the
direction is always determined to be outbound/egress.

The field options->aead_algo should be checked prior to checking above
fields. Since the same check would be required in datapath, introduce a
new flag in options for the same.

Fixes: 28dde5da503e ("app/crypto-perf: support lookaside IPsec")

Signed-off-by: Anoob Joseph <anoobj@marvell.com>
Change-Id: I63d0ae09c11125fc422911099ae0a375503f9dcb
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/92827
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>