ci: disable event timer adapter test on cn10k
Disabled event_timer_adapter_test on cn10k. This is because, one of the
cn10k board in CI is running at an SCLK of 1100 and is triggering a HW
bug. This test will be re-enabled after moving the test to a new stage
which will run only on boards with an SCLK of 1000.

Signed-off-by: Shijith Thotton <sthotton@marvell.com>
Change-Id: Ie9e719c3affe028f79eec41e919e87283cb8fb60
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/94294
Reviewed-by: Ashwin Sekhar T K <asekhar@marvell.com>
Tested-by: Ashwin Sekhar T K <asekhar@marvell.com>