examples/performance-thread: suppress compilation warning
GCC 12 throws maybe-uninitialized warning in pthread_setspecific function,
for the data argument that is const qualified. GCC is unsure if constant
pointer is iniliazed or not. Also, pthread_setspecific manual does not
insist if it can not be NULL. So, it is undefined behaviour for caller
to pass NULL data and only caller can check on it. Suppressing this warning
for now.

Signed-off-by: Gowrishankar Muthukrishnan <gmuthukrishn@marvell.com>
Change-Id: I6b3828aae0f2ef4106348e5fc77883ed93fcd9cb
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/99429
Base-Builds: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Base-Tests: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>