app/crypto-perf: return ENOTSUP for unsupported cases
dpdk-test-crypto-perf application returns failure for all
the cases which are not supported by the device.

This patch captures rte_errno to check if the case run is
supported or not, if not supported, the application would
now return ENOTSUP which can be used in automation to
identify between failed and unsupported cases.

Signed-off-by: Akhil Goyal <gakhil@marvell.com>
Change-Id: Iaec69753f44e3f19e65a36c58abdce0080ddef89
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/101374
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/101553