common/cnxk: optimize time while configuring fc on VF
PFC configuration function is taking 8 ms due
to mailbox communication to check whether sso is
connected to RQ and whether back pressure is enabled
on each aura. To optimize this time we are updating
aura attributes in nixlf and sso_ena parameter
in RQ during write configuration and the same updated
value is accessed while configuring flow control,
reducing time to 6 ms.

Change-Id: I0354ab70ebb874da7887fc0838183c6e288cfea7
Signed-off-by: Rakesh Kudurumalla <rkudurumalla@marvell.com>