common/cnxk: allow same TC on multiple RQs
To achieve actual PFC behavior, user needs to configure
different TC on different aura so that PFC can be generated
for specific TC but same TC can also configured on multiple
RQs which has same configured aura.

In this patch, aura with same BP configuration is allowed
on multiple RQs.

Signed-off-by: Sunil Kumar Kori <skori@marvell.com>
Change-Id: Id9bdd5972165450f6925367f3ef670c05f2026fa