raw/cnxk_gpio: support multi-process mode
GPIO PMD uses a mixture of standard sysfs attributes and custom
ioctl()s to control behaviour of respective GPIOs available in
the system.

This means that each userspace application, either running as a primary
or a secondary, should be able to control a set of distinct GPIOs.

In rare cases where multiple processes need to control the same set of GPIOs
userspace application is responsible for synchronizing accesses.

Signed-off-by: Tomasz Duszynski <tduszynski@marvell.com>
Change-Id: I3f8271dba1e9dedeaced88ce524af4aa5124b379
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/109541
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>