test/mbuf: fix multiple attach of external buffer
In the mbuf_autotest, an external buffer is attached to an
mbuf, then the mbuf is cloned and then the same external
mbuf is being attached to the clone.

In debug builds, this second attach of external buffer to
the clone triggers an assert as the cloned mbuf will already
have the RTE_MBUF_F_EXTERNAL flag set.

Fix this assert by avoiding the second attach of external
buffer to the clone.

Fixes: 7b295dceea07 ("test/mbuf: add unit test cases")

Signed-off-by: Rakesh Kudurumalla <rkudurumalla@marvell.com>
Signed-off-by: Ashwin Sekhar T K <asekhar@marvell.com>
Cc: lavanyax.govindarajan@intel.com
Change-Id: I9ff272174d7f95f9078b69d12679ca50525ca0fc
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpdk/+/110963
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>