net/cnxk: generalise flow operation APIs
Flow operations can be performed on cnxk ports as well as representor
ports. Since representor ports are not cnxk ports but have eswitch as
base device underneath, special handling is required to align with base
infra. Introducing a flag to generic flow APIs to discriminate if the
operation request made on normal or representor ports.

Change-Id: I91b5035ee5e8475a67c46735aefac39090bd8fb9
Signed-off-by: Harman Kalra <hkalra@marvell.com>